pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docker-mcp-orchestrator"
version = "2.0.0"
description = "Reduce MCP token usage by 90%+ — Load Docker MCP servers on-demand"
readme = "README.md"
license = {text = "CC-BY-NC-4.0"}
requires-python = ">=3.11"
authors = [
    { name = "semenovsd" }
]
keywords = [
    "mcp",
    "docker",
    "orchestrator",
    "ai",
    "llm",
    "cursor",
    "claude",
    "token-optimization",
    "lazy-loading"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
docker-mcp-orchestrator = "mcp_orchestrator:main"

[project.urls]
Homepage = "https://github.com/semenovsd/docker-mcp-orchestrator"
Documentation = "https://github.com/semenovsd/docker-mcp-orchestrator#readme"
Repository = "https://github.com/semenovsd/docker-mcp-orchestrator"
Issues = "https://github.com/semenovsd/docker-mcp-orchestrator/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_orchestrator"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/docs",
    "/config",
    "/capabilities",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "Q"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
